Clang: stop setting DWARF2 by default for FreeBSD
Previously we forced DWARF2 in order to produce debugging information
compatible with the outdated GDB in the base system and CTF tools.
We now expect GDB from ports or LLDB from base to be used for debugging,
and ctfconvert accepts newer DWARF information, so we do not need to
produce old debug information.